+++ lib/src/stripe.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'dart:math';
 
+import 'package:flutter/cupertino.dart';
 import 'package:uni_links/uni_links.dart';
 import 'package:url_launcher/url_launcher.dart';
 
@@ -16,12 +17,14 @@ class Stripe {
   /// [stripeAccount] is the id of a stripe customer and stats with "cus_".
   /// This is a optional parameter.
   ///
-  /// [returnUrlForSca] can be used to use your own return url for
+  /// [returnUrlForSca] should be used to specify a unique return url for
   /// Strong Customer Authentication (SCA) such as 3DS, 3DS2, BankID and others.
-  /// It is recommended to use your own app specific url scheme and host.
-  Stripe(String publishableKey, {String stripeAccount, String returnUrlForSca})
+  /// It is required to use your own app specific url scheme and host. This
+  /// parameter must match your "android/app/src/main/AndroidManifest.xml"
+  /// and "ios/Runner/Info.plist" configuration.
+  Stripe(String publishableKey, {String stripeAccount, @required String returnUrlForSca})
       : api = StripeApi(publishableKey, stripeAccount: stripeAccount),
-        _returnUrlForSca = returnUrlForSca ?? "stripesdk://3ds.stripesdk.io";
+        _returnUrlForSca = returnUrlForSca ?? 'stripesdk://3ds.stripesdk.io';
 
   final StripeApi api;
   final String _returnUrlForSca;
@@ -31,7 +34,7 @@ class Stripe {
   /// Throws an [Exception] if [Stripe.init] hasn't been called previously.
   static Stripe get instance {
     if (_instance == null) {
-      throw Exception("Attempted to get singleton instance of Stripe without initialization");
+      throw Exception('Attempted to get singleton instance of Stripe without initialization');
     }
     return _instance;
   }
@@ -45,12 +48,12 @@ class Stripe {
   /// [stripeAccount] is the id of a stripe customer and stats with "cus_".
   /// This is a optional parameter.
   ///
-  /// [returnUrlForSca] can be used to use your own return url for
+  /// [returnUrlForSca] should be used to specify a unique return url for
   /// Strong Customer Authentication (SCA) such as 3DS, 3DS2, BankID and others.
-  /// It is recommended to use your own app specific url scheme and host. This
+  /// It is required to use your own app specific url scheme and host. This
   /// parameter must match your "android/app/src/main/AndroidManifest.xml"
   /// and "ios/Runner/Info.plist" configuration.
-  static void init(String publishableKey, {String stripeAccount, String returnUrlForSca}) {
+  static void init(String publishableKey, {String stripeAccount, @required String returnUrlForSca}) {
     _instance = Stripe(publishableKey, stripeAccount: stripeAccount, returnUrlForSca: returnUrlForSca);
     StripeApi.init(publishableKey, stripeAccount: stripeAccount);
   }
@@ -59,35 +62,27 @@ class Stripe {
   /// This should be set on the intent before attempting to authenticate it.
   String getReturnUrlForSca() {
     final requestId = Random.secure().nextInt(99999999);
-    return "$_returnUrlForSca?requestId=$requestId";
+    return '$_returnUrlForSca?requestId=$requestId';
   }
 
-  @Deprecated("Use `Stripe.instance.getReturnUrlForSca()` instead. Will be removed in v3.0.")
-  static String getReturnUrl() {
-    final requestId = Random.secure().nextInt(99999999);
-    return "stripesdk://3ds.stripesdk.io?requestId=$requestId";
-  }
-
-  /// Confirm a SetupIntent
+  /// Authenticate a SetupIntent
   /// https://stripe.com/docs/api/setup_intents/confirm
-  Future<Map<String, dynamic>> confirmSetupIntent(String clientSecret) async {
+  Future<Map<String, dynamic>> authenticateSetupIntent(String clientSecret) async {
     final intent = await api.confirmSetupIntent(clientSecret, data: {'return_url': getReturnUrlForSca()});
     if (intent['status'] == 'requires_action') {
-      // ignore: deprecated_member_use_from_same_package
-      return handleSetupIntent(intent['next_action']);
+      return _handleSetupIntent(intent['next_action']);
     } else {
       return Future.value(intent);
     }
   }
 
-  /// Confirm a SetupIntent with a PaymentMethod
+  /// Confirm and authenticate a SetupIntent
   /// https://stripe.com/docs/api/setup_intents/confirm
-  Future<Map<String, dynamic>> confirmSetupIntentWithPaymentMethod(String clientSecret, String paymentMethod) async {
+  Future<Map<String, dynamic>> confirmSetupIntent(String clientSecret, String paymentMethod) async {
     final intent = await api
         .confirmSetupIntent(clientSecret, data: {'return_url': getReturnUrlForSca(), 'payment_method': paymentMethod});
     if (intent['status'] == 'requires_action') {
-      // ignore: deprecated_member_use_from_same_package
-      return handleSetupIntent(intent['next_action']);
+      return _handleSetupIntent(intent['next_action']);
     } else {
       return Future.value(intent);
     }
@@ -98,11 +93,10 @@ class Stripe {
   /// https://stripe.com/docs/payments/payment-intents/android
   Future<Map<String, dynamic>> confirmPayment(String paymentIntentClientSecret, {String paymentMethodId}) async {
     final data = {'return_url': getReturnUrlForSca()};
-    if (paymentMethodId != null) data["payment_method"] = paymentMethodId;
+    if (paymentMethodId != null) data['payment_method'] = paymentMethodId;
     final paymentIntent = await api.confirmPaymentIntent(paymentIntentClientSecret, data: data);
-    if (paymentIntent['status'] == "requires_action") {
-      // ignore: deprecated_member_use_from_same_package
-      return handlePaymentIntent(paymentIntent['next_action']);
+    if (paymentIntent['status'] == 'requires_action') {
+      return authenticatePaymentWithNextAction(paymentIntent['next_action']);
     } else {
       return Future.value(paymentIntent);
     }
@@ -113,10 +107,9 @@ class Stripe {
   /// https://stripe.com/docs/payments/payment-intents/android-manual
   Future<Map<String, dynamic>> authenticatePayment(String paymentIntentClientSecret) async {
     final paymentIntent = await api.retrievePaymentIntent(paymentIntentClientSecret);
-    if (paymentIntent['status'] != "requires_action") return Future.value(paymentIntent);
+    if (paymentIntent['status'] != 'requires_action') return Future.value(paymentIntent);
     final nextAction = paymentIntent['next_action'];
-    // ignore: deprecated_member_use_from_same_package
-    return handlePaymentIntent(nextAction);
+    return authenticatePaymentWithNextAction(nextAction);
   }
 
   /// Authenticate a payment with [nextAction].
@@ -124,16 +117,8 @@ class Stripe {
   /// as it avoids the request to the Stripe API to retrieve the action.
   /// To use this, return the complete [nextAction] from your server.
   Future<Map<String, dynamic>> authenticatePaymentWithNextAction(Map nextAction) async {
-    // ignore: deprecated_member_use_from_same_package
-    return handlePaymentIntent(nextAction);
-  }
-
-  /// Launch 3DS in a new browser window.
-  /// Returns a [Future] with the Stripe PaymentIntent when the user completes or cancels authentication.
-  @Deprecated("Use [authenticatePaymentWithNextAction] instead. Will be removed in v3.0.")
-  Future<Map<String, dynamic>> handlePaymentIntent(Map action) async {
     return _authenticateIntent(
-        action,
+        nextAction,
         (uri) => api.retrievePaymentIntent(
               uri.queryParameters['payment_intent_client_secret'],
             ));
@@ -141,9 +126,7 @@ class Stripe {
 
   /// Launch 3DS in a new browser window.
   /// Returns a [Future] with the Stripe SetupIntent when the user completes or cancels authentication.
-  @Deprecated(
-      "This will be removed in v3.0. Contact the maintainer if you use this and want it to remain public.")
-  Future<Map<String, dynamic>> handleSetupIntent(Map action) async {
+  Future<Map<String, dynamic>> _handleSetupIntent(Map action) async {
     return _authenticateIntent(
         action,
         (uri) => api.retrieveSetupIntent(